Move capitalize utility to shared utils library
