+++ packages/utils/package.json
@@ -8,6 +8,7 @@
   },
   "exports": {
     "./linkUtils": "./src/linkUtils.ts",
+    "./stringUtils": "./src/stringUtils.ts",
     "./types/*": "./src/types/*.ts"
   },
   "scripts": {
