+++ packages/utils/src/stringUtils.ts
